+++ app/src/main/java/com/forrestguice/suntimeswidget/EquinoxDialog.java
@@ -326,7 +326,7 @@ public class EquinoxDialog extends BottomSheetDialogFragment
     private void updateContextMenu(Context context, PopupMenu menu, final WidgetSettings.SolsticeEquinoxMode mode, final long datetime)
     {
         Intent data = new Intent();
-        data.putExtra(WorldMapDialog.EXTRA_SHOW_DATE, datetime);
+        data.putExtra(MenuAddon.EXTRA_SHOW_DATE, datetime);
         data.putExtra("mode", mode.name());
 
         Menu m = menu.getMenu();
@@ -336,10 +336,10 @@ public class EquinoxDialog extends BottomSheetDialogFragment
 
         MenuItem addonSubmenuItem = m.findItem(R.id.addonSubMenu);
         if (addonSubmenuItem != null) {
-            List<WorldMapDialog.ActivityItemInfo> addonMenuItems = WorldMapDialog.queryAddonMenuItems(context);
+            List<MenuAddon.ActivityItemInfo> addonMenuItems = MenuAddon.queryAddonMenuItems(context);
             if (!addonMenuItems.isEmpty()) {
                 SuntimesUtils.forceActionBarIcons(addonSubmenuItem.getSubMenu());
-                WorldMapDialog.populateSubMenu(addonSubmenuItem, addonMenuItems, datetime);
+                MenuAddon.populateSubMenu(addonSubmenuItem, addonMenuItems, datetime);
             } //else addonSubmenuItem.setVisible(false);
         }
     }
@@ -367,7 +367,7 @@ public class EquinoxDialog extends BottomSheetDialogFragment
             }
 
             Intent itemData = item.getIntent();
-            long itemTime = ((itemData != null) ? itemData.getLongExtra(WorldMapDialog.EXTRA_SHOW_DATE, -1L) : -1L);
+            long itemTime = ((itemData != null) ? itemData.getLongExtra(MenuAddon.EXTRA_SHOW_DATE, -1L) : -1L);
             WidgetSettings.SolsticeEquinoxMode itemMode = (itemData != null && itemData.hasExtra("mode") ? WidgetSettings.SolsticeEquinoxMode.valueOf(itemData.getStringExtra("mode")) : null);
 
             switch (item.getItemId())
@@ -413,7 +413,7 @@ public class EquinoxDialog extends BottomSheetDialogFragment
     protected void shareItem(Context context, Intent itemData)
     {
         WidgetSettings.SolsticeEquinoxMode itemMode = (itemData != null && itemData.hasExtra("mode") ? WidgetSettings.SolsticeEquinoxMode.valueOf(itemData.getStringExtra("mode")) : null);
-        long itemMillis = itemData != null ? itemData.getLongExtra(WorldMapDialog.EXTRA_SHOW_DATE, -1L) : -1L;
+        long itemMillis = itemData != null ? itemData.getLongExtra(MenuAddon.EXTRA_SHOW_DATE, -1L) : -1L;
         if (itemMode != null && itemMillis != -1L)
         {
             Calendar itemTime = Calendar.getInstance();
